feat: [SCM-45]: Move CODE specific code to 60-code
