+++ src/modules/60-code/components/SideNav/SideNav.tsx
@@ -8,14 +8,12 @@
 import React, { useMemo } from 'react'
 import { useParams, useHistory, useRouteMatch } from 'react-router-dom'
 import { Layout } from '@harness/uicore'
-import routes from '@common/RouteDefinitions'
 import { ProjectSelector, ProjectSelectorProps } from '@projects-orgs/components/ProjectSelector/ProjectSelector'
-import type { CODEPathProps } from '@common/interfaces/RouteInterfaces'
 import { SidebarLink } from '@common/navigation/SideNav/SideNav'
 import { ModuleName } from 'framework/types/ModuleName'
 import { useStrings } from 'framework/strings'
 import { useAppStore } from 'framework/AppStore/AppStoreContext'
-import { codePathProps } from '@common/utils/routeUtils'
+import routes, { codePathProps, CODEPathProps } from '../../RouteDefinitions'
 import css from './SideNav.module.scss'
 
 export default function SCMSideNav(): React.ReactElement {
